restructures the file directory, renames a few files, moves calculate coordinates to utility function
