+++ client/src/components/Clock/ClockComponents/BarCard/BarCard.jsx
