+++ client/src/components/Clock/ClockComponents/BarCard/BarCardListDisplay.css
