+++ client/src/components/Clock/ClockComponents/Tickmarks/Tickmarks.css
