+++ client/src/components/Home/Home.jsx
@@ -1,12 +1,8 @@
 import React, { useState, useEffect } from "react";
-import YelpList from "../YelpList/YelpList";
 import Clock from "../Clock/Clock";
 import Title from "../Title/Title";
 
 const Home = () => {
-	// const [isClicked, setIsClicked] = useState(false);
-	// const [submittedZip, setSubmittedZip] = useState("");
-
 	return (
 		<>
 			<Title />
